modules/locations_grpc/requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
jsonschema==3.0.2
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.9.9
python-dateutil==2.8.1
pytz==2020.1
SQLAlchemy==1.3.19
kafka-python==2.2.11